Add RUN_NOT_FOUND error and 404 distinction documentation
- Move RunNotFoundError from runs.py to core/errors.py for consistency
- Add comprehensive ERROR_CODES.md documentation
- Document 404 error distinction (PATH_NOT_FOUND vs RESOURCE_NOT_FOUND)
- Include all resource-specific 404 variants (PROJECT, AGENT, TABLE, VECTOR, RUN)
- Add troubleshooting guide and error handling examples

Addresses Epic 9, Issue 43: 404s distinguish path vs resource errors
